PS now uint16

+++ unibus.go
@@ -24,7 +24,7 @@ func (u *Unibus) physread16(a int) int {
 	case a == 0777576:
 		return u.cpu.SR2
 	case a == 0777776:
-		return u.cpu.PS
+		return int(u.cpu.PS)
 	case a&0777770 == 0777560:
 		return u.cons.consread16(a)
 	case a&0777760 == 0777400:
@@ -91,7 +91,7 @@ func (u *Unibus) physwrite16(a, v int) {
 		default:
 			panic("invalid mode")
 		}
-		u.cpu.PS = v
+		u.cpu.PS = uint16(v)
 	} else if a == 0777546 {
 		u.LKS = v
 	} else if a == 0777572 {
